html/semantics/forms/the-input-element/button.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>input type button</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#button-state-(type=button)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form id=f1>
  <input id=i1 value="foo">
  <input type=radio id=i2 checked name=button>
  <input type=checkbox id=i3>
  <input type=button id=button name=button>
</form>

<script>
  var t = async_test("clicking on button should not submit a form"),
      button = document.getElementById('button'),
      i1 = document.getElementById('i1'),
      i2 = document.getElementById('i2'),
      i3 = document.getElementById('i3');
  
  test(function(){
    assert_false(button.willValidate);
  }, "the element is barred from constraint validation");

  document.getElementById('f1').onsubmit = t.step_func(function() {
    assert_unreached("form has been submitted");
  });

  t.step(function() {
    button.click();
  });
  t.done();

  test(function(){
    i1.value = "bar";
    button.click();
    assert_equals(i1.value, "bar");
  }, "clicking on button should not reset other form fields");

  test(function(){
    assert_true(i2.checked);
    button.click();
    assert_true(i2.checked);
  }, "clicking on button should not unchecked radio buttons");

  test(function(){
    assert_false(i3.indeterminate);
    button.click();
    assert_false(i3.indeterminate);
  }, "clicking on button should not change its indeterminate IDL attribute");
</script>
